bug fixes #1 checking if user is allowed to delete post before deleting the post, in case the button is displayed wrongly #2 You need to verify your E-mail to use the app, so you can't use none existing e-mails!

+++ app/src/main/java/com/example/social_network/watchPosts.kt
@@ -199,13 +199,15 @@ class watchPosts : AppCompatActivity() {
     }
 
     fun deletePost(view: View){
-        val post = currentPost ?: return
-        val imageUrl = post.data?.get("image")
-        val storageRef = storage.getReferenceFromUrl(imageUrl.toString())
-        storageRef.delete().addOnSuccessListener {
-            db.collection("posts").document(post.id).delete().addOnSuccessListener {
-                Toast.makeText(this, "deleted", Toast.LENGTH_SHORT).show()
-                showNewPost()
+        if(auth.currentUser?.uid == posts?.get(0)?.get("uid").toString() || auth.currentUser?.uid.toString() == "27FHEnkMGUeUOlrsmBg6EdXgsHC3"){
+            val post = currentPost ?: return
+            val imageUrl = post.data?.get("image")
+            val storageRef = storage.getReferenceFromUrl(imageUrl.toString())
+            storageRef.delete().addOnSuccessListener {
+                db.collection("posts").document(post.id).delete().addOnSuccessListener {
+                    Toast.makeText(this, "deleted", Toast.LENGTH_SHORT).show()
+                    showNewPost()
+                }
             }
         }
     }
